Added swagger and some of the items

+++ PictureBasketApi/Models/ProductType.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public string Title { get; set; }
-        public decimal Width { get; set; }
+        public double Width { get; set; }
         public int StackLimit { get; set; }
     }
 }
